Move the utils getTransactionStatus and getChainName to a new module utils/labels
